third_party/WebKit/LayoutTests/platform/win/external/wpt/mediacapture-streams/GUM-api.https-expected.txt

This is a testharness.js-based test.
PASS getUserMedia() is present on navigator 
Harness: the test ran to completion.
